# server/openapi.yaml
openapi: 3.1.0
info:
  title: WPPConnect Dev Platform API
  version: 1.0.0
  description: Programmable WhatsApp sessions, messaging, media, contacts, statuses, and webhooks.
servers:
  - url: http://localhost:4005/api/v1
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
    TextMessage:
      type: object
      required: [to, text]
      properties:
        to: { type: string, examples: [15551234567] }
        text: { type: string }
    FileMessage:
      type: object
      required: [to, dataUrl, filename]
      properties:
        to: { type: string }
        dataUrl: { type: string, description: Data URL containing base64 file data }
        filename: { type: string }
        caption: { type: string }
    LocationMessage:
      type: object
      required: [to, latitude, longitude]
      properties:
        to: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        title: { type: string }
security:
  - ApiKey: []
paths:
...
  /chats/{chatId}/messages:
    get:
      summary: Read recent messages
      parameters:
        - { in: path, name: chatId, required: true, schema: { type: string } }
        - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 100, default: 30 } }
      responses: { '200': { description: Messages } }
...
  /messages/text:
    post:
      summary: Send a text message
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TextMessage' } } } }
      responses: { '201': { description: Message created } }
...
  /messages/file:
    post:
      summary: Send a file, image, video, audio, or voice note
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/FileMessage' } } } }
      responses: { '201': { description: Message created } }
...
  /messages/location:
    post:
      summary: Send a location
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/LocationMessage' } } } }
      responses: { '201': { description: Location message created } }
...
  /webhooks/{id}:
    delete:
      summary: Delete a webhook subscription
      parameters: [{ in: path, name: id, required: true, schema: { type: string } }]
      responses: { '200': { description: Webhook deleted } }